server/requirements.txt
paho-mqtt==1.6.1
influxdb-client==1.36.1
Flask==2.3.3